api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
squareup
httpx
SQLAlchemy>=2.0
asyncpg